script/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "summaryarena"
version = "0.1.0"
description = "CLI tool to benchmark LLM summarization quality across providers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Summary Arena Contributors" },
]
keywords = ["llm", "summarization", "benchmark", "arena"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer[all]>=0.9.0",
    "litellm>=1.30.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
metrics = [
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
]
curation = [
    "datasets>=2.14.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
all = [
    "summaryarena[metrics,curation,dev]",
]

[project.scripts]
summaryarena = "summaryarena.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["summaryarena*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
